+++ CHANGELOG.md
@@ -27,6 +27,10 @@ This release changes the pinned API version to 2026-06-24.preview.
   * Add support for event notification `V2MoneyManagementOutboundTransferUnderReviewEvent` with related object `V2::MoneyManagement::OutboundTransfer`
   * ⚠️ Remove support for event notifications `V2CoreAccountIncludingConfigurationStorerCapabilityStatusUpdatedEvent` and `V2CoreAccountIncludingConfigurationStorerUpdatedEvent` with related object `V2::Core::Account`
 
+## 19.3.1 - 2026-07-15
+* [#1902](https://github.com/stripe/stripe-ruby/pull/1902) Replace source hash with Telemetry UUID
+* [#1901](https://github.com/stripe/stripe-ruby/pull/1901) Make Error fields generated
+
 ## 19.3.0 - 2026-06-24
 This release changes the pinned API version to 2026-06-24.dahlia.